Leverage distribution/reference for local index
This has 2 upside:
- it enforces bundle names in the local index are valid repo names
- it uses a proven image reference parser instead of relying on raw strings package

+++ cmd/duffle/bundle_remove.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -10,6 +11,7 @@ import (
 	"github.com/deis/duffle/pkg/repo"
 
 	"github.com/Masterminds/semver"
+	"github.com/docker/distribution/reference"
 	"github.com/spf13/cobra"
 )
 
@@ -31,7 +33,16 @@ func newBundleRemoveCmd(w io.Writer) *cobra.Command {
 		Long:    bundleRemoveDesc,
 		Args:    cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			bname := args[0]
+			named, err := reference.ParseNormalizedNamed(args[0])
+			if err != nil {
+				return err
+			}
+			tagged, _ := named.(reference.NamedTagged)
+			if tagged != nil {
+				if versions != "" {
+					return errors.New("cannot set --version flag when bundle name has a tag")
+				}
+			}
 
 			h := home.Home(homePath())
 			index, err := repo.LoadIndex(h.Repositories())
@@ -39,9 +50,9 @@ func newBundleRemoveCmd(w io.Writer) *cobra.Command {
 				return err
 			}
 
-			vers, ok := index.GetVersions(bname)
+			vers, ok := index.GetVersions(named)
 			if !ok {
-				fmt.Fprintf(w, "Bundle %q not found. Nothing deleted.", bname)
+				fmt.Fprintf(w, "Bundle %q not found. Nothing deleted.", args[0])
 				return nil
 			}
 
@@ -61,10 +72,14 @@ func newBundleRemoveCmd(w io.Writer) *cobra.Command {
 					if ok, _ := matcher.Validate(sv); ok {
 						fmt.Fprintf(w, "Version %s matches constraint %q\n", ver, versions)
 						deletions[ver] = sha
-						index.DeleteVersion(bname, ver)
+						versioned, err := reference.WithTag(named, ver)
+						if err != nil {
+							return err
+						}
+						index.DeleteVersion(versioned)
 						// If there are no more versions, remove the entire entry.
-						if vers, ok := index.GetVersions(bname); ok && len(vers) == 0 {
-							index.Delete(bname)
+						if vers, ok := index.GetVersions(named); ok && len(vers) == 0 {
+							index.DeleteAll(named)
 						}
 
 					}
@@ -79,10 +94,27 @@ func newBundleRemoveCmd(w io.Writer) *cobra.Command {
 				deleteBundleVersions(deletions, index, h, w)
 				return nil
 			}
+			if tagged != nil {
+				sha, ok := vers[tagged.Tag()]
+				if !ok {
+					return fmt.Errorf("version %q not found", tagged.Tag())
+				}
+				index.DeleteVersion(tagged)
+				if err := index.WriteFile(h.Repositories(), 0644); err != nil {
+					return err
+				}
+				if !isShaReferenced(index, sha) {
+					fpath := filepath.Join(h.Bundles(), sha)
+					if err := os.Remove(fpath); err != nil {
+						fmt.Fprintf(w, "WARNING: could not delete stake record %q", fpath)
+					}
+				}
+				return nil
+			}
 
 			// If no version was specified, delete entire record
-			if !index.Delete(bname) {
-				fmt.Fprintf(w, "Bundle %q not found. Nothing deleted.", bname)
+			if !index.DeleteAll(named) {
+				fmt.Fprintf(w, "Bundle %q not found. Nothing deleted.", named)
 				return nil
 			}
 			if err := index.WriteFile(h.Repositories(), 0644); err != nil {
@@ -98,14 +130,27 @@ func newBundleRemoveCmd(w io.Writer) *cobra.Command {
 	return cmd
 }
 
+func isShaReferenced(index repo.Index, sha string) bool {
+	for _, vs := range index {
+		for _, otherSha := range vs {
+			if otherSha == sha {
+				return true
+			}
+		}
+	}
+	return false
+}
+
 // deleteBundleVersions removes the given SHAs from bundle storage
 //
 // It warns, but does not fail, if a given SHA is not found.
 func deleteBundleVersions(vers map[string]string, index repo.Index, h home.Home, w io.Writer) {
 	for _, sha := range vers {
-		fpath := filepath.Join(h.Bundles(), sha)
-		if err := os.Remove(fpath); err != nil {
-			fmt.Fprintf(w, "WARNING: could not delete stake record %q", fpath)
+		if !isShaReferenced(index, sha) {
+			fpath := filepath.Join(h.Bundles(), sha)
+			if err := os.Remove(fpath); err != nil {
+				fmt.Fprintf(w, "WARNING: could not delete stake record %q", fpath)
+			}
 		}
 	}
 }
